+++ languagetool-core/src/main/java/org/languagetool/markup/AnnotatedText.java
@@ -86,7 +86,9 @@ public class AnnotatedText {
   public String getTextWithMarkup() {
     StringBuilder sb = new StringBuilder();
     for (TextPart part : parts) {
-      sb.append(part.getPart());
+      if (part.getType() != TextPart.Type.FAKE_CONTENT) {
+        sb.append(part.getPart());
+      }
     }
     return sb.toString();
   }