+++ languagetool-language-modules/es/src/main/java/org/languagetool/tokenizers/es/SpanishWordTokenizer.java
@@ -31,9 +31,6 @@ import org.languagetool.tokenizers.WordTokenizer;
  */
 public class SpanishWordTokenizer extends WordTokenizer {
 
-  public SpanishWordTokenizer() {
-  }
-
   @Override
   public List<String> tokenize(String text) {
     List<String> tokenList = new ArrayList<>();
@@ -45,7 +42,7 @@ public class SpanishWordTokenizer extends WordTokenizer {
                     + "\u2028\u2029\u202a\u202b\u202c\u202d\u202e\u202f"
                     + "\u205F\u2060\u2061\u2062\u2063\u206A\u206b\u206c\u206d"
                     + "\u206E\u206F\u3000\u3164\ufeff\uffa0\ufff9\ufffa\ufffb"
-                    + ",.;()[]{}<>!?:/\\\"'«»„”“‘`’…¿¡\t\n\r", true);
+                    + ",.;()[]{}<>!?:=*#∗×·+÷/\\\"'«»„”“‘`’…¿¡\t\n\r", true);
     while (st.hasMoreElements()) {
       tokenList.add(st.nextToken());
     }